refactor to package 'decorrelategrids'
